Module Dkml_package_console_common.Author_types

type build_info = {

package_name : string;

}

The type of build information.

package_name - Which opam package the program will be generated from

type program_name = {

name_full : string;

name_camel_case_nospaces : string;

name_kebab_lower_case : string;

installation_prefix_camel_case_nospaces_opt : string option;

installation_prefix_kebab_lower_case_opt : string option;

}

The type of program names.

name_full - Examples include "Diskuv OCaml"

name_camel_case_nospaces - If the program name was "Diskuv OCaml" then the name_camel_case_nospaces could be either "DiskuvOCaml" or "DiskuvOcaml".

name_kebab_lower_case - If the program name was "Diskuv OCaml" then the name_kebab_lower_case would be "diskuv-ocaml".

installation_prefix_camel_case_nospaces_opt - The name used when constructing an installation prefix that takes a CamelCase with no spaces. If not specified, then name_camel_case_nospaces is used.

installation_prefix_kebab_lower_case_opt - The name used when constructing an installation prefix that takes a kebab-lower-case. If not specified, then name_kebab_lower_case is used.

type organization = {

legal_name : string;

common_name_full : string;

common_name_camel_case_nospaces : string;

common_name_kebab_lower_case : string;

}

Details about the organization for signing binaries.

legal_name - Examples include "Diskuv, Inc."

common_name_full - Examples include "Dow Jones"

common_name_camel_case_nospaces - If the common name was "Dow Jones" then the common_name_camel_case_nospaces would be "DowJones".

common_name_kebab_lower_case - If the program name was "Dow Jones" then the common_name_kebab_lower_case would be "dow-jones".

type program_assets = {

logo_icon_32x32_opt : string option;

}

Logos and other assets used during installation.

logo_icon_32x32_opt - None or the Some icon_data where icon_data is the contents of the 32x32 icon file. Do not use a filename or a URL; instead read the file into a string. Using the ocaml-crunch package will automate this for you.

type program_info = {

url_info_about_opt : string option;

url_update_info_opt : string option;

help_link_opt : string option;

estimated_byte_size_opt : int64 option;

windows_language_code_id_opt : int option;

embeds_32bit_uninstaller : bool;

embeds_64bit_uninstaller : bool;

}

Information about the program.

url_info_about_opt - A URL to a webpage describing the program.

url_update_info_opt - A URL to a webpage describing updates to the program.

help_link_opt - A URL to a help webpage.

estimated_byte_size_opt - How much disk space, in bytes, is estimated to be taken away after the installation is complete.

windows_language_code_id_opt - The Language ID for a Windows installation from https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-lcid/70feba9f-294e-491e-b6eb-56532684c37f. Defaults to 0x0409 (en-US).

A simpler non-exhaustive list is at https://docs.microsoft.com/en-us/openspecs/office_standards/ms-oe376/6c085406-a698-4e12-9d4d-c3b0ee3dbc4a