Module Dkml_package_console_common.Windows_registry

val delete_program_entry : program_name:Author_types.program_name -> unit Dkml_install_api.Forward_progress.t

Delete from the Windows registry so that the program will not appear in Add/Remove Programs

val write_program_entry : installation_prefix:Fpath.t -> organization:Author_types.organization -> program_name:Author_types.program_name -> program_assets:Author_types.program_assets -> program_version:string -> program_info:Author_types.program_info -> unit Dkml_install_api.Forward_progress.t

Write into the Windows registry so that the program will appear Add/Remove Programs